refactor(paywalls-v2): address haptic feedback review (structure/naming)
- Move SelectionHapticFeedback into its own file (was in PaywallEventTracker).
- Rename PackageSelectionHapticFeedback -> SelectionHapticFeedback and the
  env value packageSelectionHapticFeedback -> selectionHapticFeedback (it's
  used for tab switches too, not just packages).
- Centralize the `?? true` default via a computed on the tab views.
- Keep the warm-up rationale on prepare() only, drop the repeated comments.
