+++ RevenueCatUI/Templates/V2/Components/Tabs/TabControlToggleComponentView.swift
@@ -38,7 +38,7 @@ struct TabControlToggleComponentView: View {
     @Environment(\.componentInteractionLogger)
     private var componentInteractionLogger
 
-    @Environment(\.packageSelectionHapticFeedback)
+    @Environment(\.selectionHapticFeedback)
     private var hapticFeedback
 
     private let viewModel: TabControlToggleComponentViewModel
@@ -63,11 +63,15 @@ struct TabControlToggleComponentView: View {
                     componentName: self.tabControlContext.name,
                     isOn: newValue
                 ))
-                if self.viewModel.component.hapticFeedbackEnabled ?? true {
+                if self.hapticFeedbackEnabled {
                     self.hapticFeedback()
                 }
             }
         )
+    }
+
+    private var hapticFeedbackEnabled: Bool {
+        self.viewModel.component.hapticFeedbackEnabled ?? true
     }
 
     init(viewModel: TabControlToggleComponentViewModel, onDismiss: @escaping () -> Void) {
@@ -86,10 +90,8 @@ struct TabControlToggleComponentView: View {
                 )
             )
             .labelsHidden()
-            // Warm the haptics engine as the toggle appears, so the first toggle's
-            // render isn't stalled behind the one-time engine load.
             .onAppear {
-                if self.viewModel.component.hapticFeedbackEnabled ?? true {
+                if self.hapticFeedbackEnabled {
                     self.hapticFeedback.prepare()
                 }
             }
